+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -1,6 +1,7 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
 import insane96mcp.iguanatweaksreborn.event.SREventFactory;
+import insane96mcp.iguanatweaksreborn.module.combat.stats.Stats;
 import insane96mcp.iguanatweaksreborn.module.experience.PlayerExperience;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import net.minecraft.core.BlockPos;
@@ -13,9 +14,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import javax.annotation.Nullable;
@@ -66,5 +65,15 @@ public abstract class PlayerMixin extends LivingEntity {
 	@ModifyVariable(method = "actuallyHurt", at = @At(value = "STORE", ordinal = 2), argsOnly = true, ordinal = 0)
 	public float onPreAbsorptionCalculation(float amount, DamageSource damageSource) {
 		return SREventFactory.onLivingHurtPreAbsorption(this, damageSource, amount);
+	}
+
+	@ModifyConstant(method = "attack", constant = @Constant(floatValue = 0.2f, ordinal = 0))
+	public float attackStrengthAtMaxCooldown(float value) {
+        return Stats.noDamageWhenSpamming() ? 0f : value;
+    }
+
+	@ModifyConstant(method = "attack", constant = @Constant(floatValue = 0.8f, ordinal = 0))
+	public float attackStrengthAtFullSwing(float value) {
+		return Stats.noDamageWhenSpamming() ? 1f : value;
 	}
 }