+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/stats/Stats.java
@@ -53,6 +53,9 @@ public class Stats extends Feature {
 	@Label(name = "Player attack range modifier", description = "Adds this to players' attack range")
 	public static Double playerAttackRangeModifier = -0.5d;
 	@Config
+	@Label(name = "Player no damage when spamming", description = "In vanilla, if you attack as soon as you just attacked you already deal 20% of the full damage. This changes that to 0%.")
+	public static Boolean playerNoDamageWhenSpamming = true;
+	@Config
 	@Label(name = "Players movement speed reduction", description = "Reduces movement speed for players by this percentage.")
 	public static Double playersMovementSpeedReduction = 0.05d;
 	@Config
@@ -97,6 +100,10 @@ public class Stats extends Feature {
 			MobEffects.WEAKNESS.addAttributeModifier(Attributes.ATTACK_DAMAGE, "22653B89-116E-49DC-9B6B-9971489B5BE5", 0.0D, AttributeModifier.Operation.MULTIPLY_BASE);
 			((AttackDamageMobEffect)MobEffects.WEAKNESS).multiplier = -0.2d;
 		}
+	}
+
+	public static boolean noDamageWhenSpamming() {
+		return isEnabled(Stats.class) && playerNoDamageWhenSpamming;
 	}
 
 	@SubscribeEvent